pyproject.toml
[project]
name = "dndproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "bcrypt==5.0.0",
    "certifi==2026.1.4",
    "click==8.3.1",
    "ecdsa==0.19.1",
    "fastapi==0.128.0",
    "greenlet==3.3.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "packaging==25.0",
    "playwright==1.55.0",
    "psycopg[binary]>=3.2.0",
    "pyasn1==0.6.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pyee==13.0.0",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.21",
    "rsa==4.9.1",
    "six==1.17.0",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "tenacity==9.1.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.40.0",
]
